requirements.txt
matplotlib==3.5.1
numpy==1.22.3
opencv_python==4.5.1.48
pandas==1.4.2
Pillow==9.3.0
planar==0.4
pydicom==2.2.2
scikit_image==0.18.1
skimage==0.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
